requirements.txt
aiofiles==24.1.0
aiogram==3.15.0
aiogram_i18n==1.4
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
asyncio==3.4.3
asyncpg==0.30.0
attrs==24.2.0
babel==2.16.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
contrib==0.3.0
cycler==0.12.1
Flask==3.1.0
flask-ngrok==0.0.25
fonttools==4.55.3
frozenlist==1.5.0
google-api-core==2.24.0
google-api-python-client==2.158.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
greenlet==3.1.1
gspread==6.0.2
gspread_asyncio==2.0.0
gunicorn==23.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httplib2==0.22.0
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
magic-filter==1.0.12
MarkupSafe==3.0.2
matplotlib==3.10.0
multidict==6.1.0
ngrok==1.4.0
numpy==2.2.1
oauth2client==4.1.3
oauthlib==3.2.2
packaging==24.2
pillow==11.0.0
priority==2.0.0
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.3
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pyngrok==7.2.2
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
Quart==0.19.9
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.22.3
rsa==4.9
six==1.17.0
SQLAlchemy==2.0.36
StrEnum==0.4.15
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.18.3
